real-check/pyproject.toml
[project]
name = "real-check"
version = "0.1.0"
description = "Live USB checks for AlphaSmart NEO direct USB communication"
authors = [
    { name = "Jakub Kolcar", email = "j.kolcar@raynet.cz" }
]
requires-python = ">=3.12"
dependencies = [
    "pyusb>=1.2.1",
    "neotools",
]

[project.scripts]
real-check = "real_check:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/real_check"]

[tool.uv.sources]
neotools = { path = "../poc/neotools" }
